(PC-17986) chore(Home): create a modules folder and move modules component to clean the home/component folder
