+++ src/features/home/components/index.ts
@@ -1,5 +1,5 @@
-export { BusinessModule } from './BusinessModule'
-export { ExclusivityModule } from './ExclusivityModule'
 export { HomeHeader } from './HomeHeader'
-export { OffersModule } from './OffersModule'
-export { VenuesModule } from './VenuesModule'
+export { BusinessModule } from './modules/BusinessModule'
+export { ExclusivityModule } from './modules/ExclusivityModule'
+export { OffersModule } from './modules/OffersModule'
+export { VenuesModule } from './modules/VenuesModule'
