+++ src/features/home/components/modules/ExclusivityModule.utils.test.ts
@@ -1,4 +1,4 @@
-import { shouldDisplayExcluOffer } from 'features/home/components/ExclusivityModule.utils'
+import { shouldDisplayExcluOffer } from 'features/home/components/modules/ExclusivityModule.utils'
 import { ExclusivityPane } from 'features/home/contentful'
 import { offerResponseSnap as offer } from 'features/offer/api/snaps/offerResponseSnap'
 import { GeoCoordinates } from 'libs/geolocation'
