+++ src/features/home/components/modules/OffersModule.tsx
